requirements.txt
aiofiles==23.2.1 ; python_version >= "3.8" and python_version < "3.12"
altair==5.3.0 ; python_version >= "3.8" and python_version < "3.12"
annotated-types==0.6.0 ; python_version >= "3.8" and python_version < "3.12"
anyio==4.3.0 ; python_version >= "3.8" and python_version < "3.12"
attrs==23.2.0 ; python_version >= "3.8" and python_version < "3.12"
certifi==2024.2.2 ; python_version >= "3.8" and python_version < "3.12"
charset-normalizer==3.3.2 ; python_version >= "3.8" and python_version < "3.12"
click==8.1.7 ; python_version >= "3.8" and python_version < "3.12"
colorama==0.4.6 ; python_version >= "3.8" and python_version < "3.12" and (platform_system == "Windows" or sys_platform == "win32")
contourpy==1.1.1 ; python_version >= "3.8" and python_version < "3.12"
cycler==0.12.1 ; python_version >= "3.8" and python_version < "3.12"
dnspython==2.6.1 ; python_version >= "3.8" and python_version < "3.12"
email-validator==2.1.1 ; python_version >= "3.8" and python_version < "3.12"
exceptiongroup==1.2.1 ; python_version >= "3.8" and python_version < "3.11"
fastapi-cli==0.0.3 ; python_version >= "3.8" and python_version < "3.12"
fastapi==0.111.0 ; python_version >= "3.8" and python_version < "3.12"
ffmpy==0.3.2 ; python_version >= "3.8" and python_version < "3.12"
filelock==3.14.0 ; python_version >= "3.8" and python_version < "3.12"
fonttools==4.51.0 ; python_version >= "3.8" and python_version < "3.12"
fsspec==2024.5.0 ; python_version >= "3.8" and python_version < "3.12"
gradio-client==0.16.3 ; python_version >= "3.8" and python_version < "3.12"
gradio-highlightedtextbox==0.0.12 ; python_version >= "3.8" and python_version < "3.12"
gradio==4.31.3 ; python_version >= "3.8" and python_version < "3.12"
h11==0.14.0 ; python_version >= "3.8" and python_version < "3.12"
httpcore==1.0.5 ; python_version >= "3.8" and python_version < "3.12"
httptools==0.6.1 ; python_version >= "3.8" and python_version < "3.12"
httpx==0.27.0 ; python_version >= "3.8" and python_version < "3.12"
huggingface-hub==0.23.0 ; python_version >= "3.8" and python_version < "3.12"
idna==3.7 ; python_version >= "3.8" and python_version < "3.12"
importlib-resources==6.4.0 ; python_version >= "3.8" and python_version < "3.12"
jinja2==3.1.4 ; python_version >= "3.8" and python_version < "3.12"
jsonschema-specifications==2023.12.1 ; python_version >= "3.8" and python_version < "3.12"
jsonschema==4.22.0 ; python_version >= "3.8" and python_version < "3.12"
kiwisolver==1.4.5 ; python_version >= "3.8" and python_version < "3.12"
markdown-it-py==3.0.0 ; python_version >= "3.8" and python_version < "3.12"
markupsafe==2.1.5 ; python_version >= "3.8" and python_version < "3.12"
matplotlib==3.7.5 ; python_version >= "3.8" and python_version < "3.12"
mdurl==0.1.2 ; python_version >= "3.8" and python_version < "3.12"
numpy==1.24.4 ; python_version >= "3.8" and python_version < "3.12"
orjson==3.10.3 ; python_version >= "3.8" and python_version < "3.12"
packaging==24.0 ; python_version >= "3.8" and python_version < "3.12"
pandas==2.0.3 ; python_version >= "3.8" and python_version < "3.12"
pillow==10.3.0 ; python_version >= "3.8" and python_version < "3.12"
pkgutil-resolve-name==1.3.10 ; python_version >= "3.8" and python_version < "3.9"
pydantic-core==2.18.2 ; python_version >= "3.8" and python_version < "3.12"
pydantic==2.7.1 ; python_version >= "3.8" and python_version < "3.12"
pydub==0.25.1 ; python_version >= "3.8" and python_version < "3.12"
pygments==2.18.0 ; python_version >= "3.8" and python_version < "3.12"
pyparsing==3.1.2 ; python_version >= "3.8" and python_version < "3.12"
python-dateutil==2.9.0.post0 ; python_version >= "3.8" and python_version < "3.12"
python-dotenv==1.0.1 ; python_version >= "3.8" and python_version < "3.12"
python-multipart==0.0.9 ; python_version >= "3.8" and python_version < "3.12"
pytz==2024.1 ; python_version >= "3.8" and python_version < "3.12"
pyyaml==6.0.1 ; python_version >= "3.8" and python_version < "3.12"
referencing==0.35.1 ; python_version >= "3.8" and python_version < "3.12"
requests==2.31.0 ; python_version >= "3.8" and python_version < "3.12"
rich==13.7.1 ; python_version >= "3.8" and python_version < "3.12"
rpds-py==0.18.1 ; python_version >= "3.8" and python_version < "3.12"
ruff==0.2.2 ; python_version >= "3.8" and python_version < "3.12" and sys_platform != "emscripten"
semantic-version==2.10.0 ; python_version >= "3.8" and python_version < "3.12"
shellingham==1.5.4 ; python_version >= "3.8" and python_version < "3.12"
six==1.16.0 ; python_version >= "3.8" and python_version < "3.12"
sniffio==1.3.1 ; python_version >= "3.8" and python_version < "3.12"
starlette==0.37.2 ; python_version >= "3.8" and python_version < "3.12"
toml==0.10.2 ; python_version >= "3.8" and python_version < "3.12"
tomlkit==0.12.0 ; python_version >= "3.8" and python_version < "3.12"
toolz==0.12.1 ; python_version >= "3.8" and python_version < "3.12"
tqdm==4.66.4 ; python_version >= "3.8" and python_version < "3.12"
typer==0.12.3 ; python_version >= "3.8" and python_version < "3.12"
typing-extensions==4.11.0 ; python_version >= "3.8" and python_version < "3.12"
tzdata==2024.1 ; python_version >= "3.8" and python_version < "3.12"
ujson==5.10.0 ; python_version >= "3.8" and python_version < "3.12"
urllib3==2.2.1 ; python_version >= "3.8" and python_version < "3.12"
uvicorn==0.29.0 ; python_version >= "3.8" and python_version < "3.12" and sys_platform != "emscripten"
uvicorn[standard]==0.29.0 ; python_version >= "3.8" and python_version < "3.12"
uvloop==0.19.0 ; (sys_platform != "win32" and sys_platform != "cygwin") and platform_python_implementation != "PyPy" and python_version >= "3.8" and python_version < "3.12"
watchfiles==0.21.0 ; python_version >= "3.8" and python_version < "3.12"
websockets==11.0.3 ; python_version >= "3.8" and python_version < "3.12"
zipp==3.18.2 ; python_version >= "3.8" and python_version < "3.10"
